requirements.txt
fastapi
pydantic
uvicorn
python-dotenv
requests
beautifulsoup4
feedparser
